requirements.txt
# --- Web Server ---
Flask==3.1.2
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.5.1
gunicorn
eventlet

# --- Database ---
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pgvector==0.3.6

# --- LangChain / RAG ---
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langsmith==0.4.31
scikit-learn

# --- LLM Provider ---
openai==1.109.1
tiktoken==0.11.0

# --- HTTP ---
httpx==0.28.1
requests==2.32.5

# --- Config ---
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0

# --- Utilities ---
orjson==3.11.3
tenacity==9.1.2
tqdm==4.67.1
regex==2025.9.18
PyYAML==6.0.3
pillow==11.3.0
WTForms==3.2.1
